⏪️ revert(api): change the `CookieService` directory to modules
