+++ apps/api/src/app/auth/[client]/login/[provider]/callback.ts
@@ -14,9 +14,9 @@ import { OAuthProviderGateway } from "../../../../../interface-adapter/gateway/o
 import { OAuthAccountRepository } from "../../../../../interface-adapter/repositories/oauth-account";
 import { SessionRepository } from "../../../../../interface-adapter/repositories/session";
 import { oAuthProviderSchema } from "../../../../../models/entities/oauth-account";
+import { CookieService } from "../../../../../modules/cookie";
 import { ElysiaWithEnv } from "../../../../../modules/elysia-with-env";
 import { rateLimiter } from "../../../../../modules/rate-limiter";
-import { CookieService } from "../../../../../services/cookie";
 import { SessionTokenService } from "../../../../../services/session-token";
 
 const cookieSchemaObject = {
