+++ apps/api/src/app/auth/[client]/login/[provider]/index.ts
@@ -9,9 +9,9 @@ import {
 import { clientSchema } from "../../../../../common/schema";
 import { OAuthProviderGateway } from "../../../../../interface-adapter/gateway/oauth-provider";
 import { oAuthProviderSchema } from "../../../../../models/entities/oauth-account";
+import { CookieService } from "../../../../../modules/cookie";
 import { ElysiaWithEnv } from "../../../../../modules/elysia-with-env";
 import { rateLimiter } from "../../../../../modules/rate-limiter";
-import { CookieService } from "../../../../../services/cookie";
 import { ProviderCallback } from "./callback";
 
 const cookieSchemaObject = {
