+++ apps/api/src/app/auth/[client]/signup/index.ts
@@ -6,9 +6,9 @@ import { DrizzleService } from "../../../../infrastructure/drizzle";
 import { SessionRepository } from "../../../../interface-adapter/repositories/session";
 import { UserRepository } from "../../../../interface-adapter/repositories/user";
 import { UserCredentialRepository } from "../../../../interface-adapter/repositories/user-credential";
+import { CookieService } from "../../../../modules/cookie";
 import { ElysiaWithEnv } from "../../../../modules/elysia-with-env";
 import { rateLimiter } from "../../../../modules/rate-limiter";
-import { CookieService } from "../../../../services/cookie";
 import { PasswordService } from "../../../../services/password";
 import { SessionTokenService } from "../../../../services/session-token";
 import { Provider } from "./[provider]";
