+++ apps/api/src/modules/cookie/index.ts
@@ -1,0 +1,2 @@
+export { CookieService } from "./cookie.service";
+export type { ICookieService, CookieAttributes } from "./interfaces/cookie.service.interface";
