--- apps/api/src/services/cookie/index.ts
@@ -1,3 +1,0 @@
-export { CookieService } from "./cookie.service";
-export type { ICookieService } from "./interfaces/cookie.service.interface";
-export type { CookieAttributes } from "./interfaces/cookie.type";
