--- apps/api/src/services/cookie/interfaces/cookie.type.ts
@@ -1,9 +1,0 @@
-export type CookieAttributes = {
-	secure?: boolean;
-	path?: string;
-	domain?: string;
-	sameSite?: "lax" | "strict" | "none";
-	httpOnly?: boolean;
-	maxAge?: number;
-	expires?: Date;
-};
